requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.4
django-environ==0.11.2
djangorestframework==3.15.2
et_xmlfile==2.0.0
idna==3.10
kagglehub==0.3.6
mysql-connector-python==9.1.0
numpy==2.2.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
tzdata==2024.2
urllib3==2.3.0
xlrd==2.0.1
